Fix __init__ directory change for compiled scrip
Determining the script's working directory using the previous
method doesn't work when the script is compiled with PyInstaller.
Had to use a different method.

+++ ocvbot/__init__.py
@@ -1,3 +1,4 @@
+# coding=UTF-8
 """
 Sets up global variables and a few other preparatory checks.
 
@@ -8,21 +9,20 @@ import sys
 
 import pyautogui as pag
 
+pag.PAUSE = 0
 sys.setrecursionlimit(9999)
-
 log.basicConfig(format='%(asctime)s %(filename)s.%(funcName)s - %(message)s',
                 level='INFO')
 
-pag.PAUSE = 0
-
-# TODO: Fix this for PyInstaller.
 # Make sure the program's working directory is the directory in which
-#   this file is located.
-absolute_path = os.path.abspath(__file__)
-dir_name = os.path.dirname(absolute_path)
-os.chdir(dir_name)
+#   this file is located. If the script is compiled (i.e. "frozen"), a
+#   different method must be used.
+if hasattr(sys, "frozen"):
+    os.chdir(os.path.dirname(sys.executable))
+else:
+    os.chdir(os.path.dirname(__file__))
 
 # TODO: Find a better way to do this.
-# TODO: Add a check to only run this if the OS is Linux.
 # Clean up left over screenshots from previous runs.
-os.system('rm .screenshot2*.png >/dev/null 2>&1')
+if os.name == 'posix':
+    os.system('rm .screenshot2*.png >/dev/null 2>&1')
